move all components related to books to their own folder
