+++ assets/src/components/books/detail/BookDetail.jsx
