+++ assets/src/components/books/detail/BookPublicationInfo.jsx
