+++ assets/src/components/books/detail/BookPublicationInfo.test.jsx
