+++ assets/src/components/mybooks/MyBooks.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import Paper from '@material-ui/core/Paper';
 import Icon from '@material-ui/core/Icon';
 import { getMyBooks } from '../../services/BookService';
-import BookList from '../libraries/BookList';
+import BookList from '../books/BookList';
 
 export default class MyBooks extends Component {
   constructor(props) {
